+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix choice disable state wasn't considered when showing the "no choices to choose from" notice
 * Fix regression "no choices to choose from" notice not triggering when no selectable choices exist for select-one. [#1185](https://github.com/Choices-js/Choices/issues/1185)
 * Fix regression where webpack doesn't permit importing scss/css @tagliala (#1193)
+* Fix regression of UnhighlightItem event not firing [#1173](https://github.com/Choices-js/Choices/issues/1173)
 
 ### Chore
 * Add e2e tests for "no choices" behavior to match v10