+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Fix choice disable state wasn't considered when showing the "no choices to choose from" notice
 * Fix regression where webpack doesn't permit importing scss/css @tagliala [#1193](https://github.com/Choices-js/Choices/issues/1193)
 * Fix regression "no choices to choose from"/"no results found" notice did not reliably trigger. [#1185](https://github.com/Choices-js/Choices/issues/1185) [#1191](https://github.com/Choices-js/Choices/issues/1191)
+* Fix regression of UnhighlightItem event not firing [#1173](https://github.com/Choices-js/Choices/issues/1173)
 
 ### Chore
 * Add e2e tests for "no choices" behavior to match v10